[chore] ViewModel 분리
