+++ SMASHING/Presentation/Address/AddressSearchController.swift
@@ -6,96 +6,6 @@
 //
 
 import Combine
-
-protocol AddressSearchViewModelProtocol: InputOutputProtocol where Input == AddressSearchViewModel.Input, Output == AddressSearchViewModel.Output {
-    var searchResults: [String] {get}
-    associatedtype NavigationEvent
-    var navigationEvent: NavigationEvent {get}
-}
-
-final class AddressSearchViewModel: AddressSearchViewModelProtocol {
-    
-    var searchResults: [String] = []
-    var searchCurrentPage: Int = 1
-    private var isFetching = false
-    private var currentQuery: String = ""
-    private var cancellables: Set<AnyCancellable> = []
-    
-    enum Input {
-        case searchAddress(String)
-        case reachedBottom
-        case addressSelected(String)
-        case backTapped
-    }
-    
-    struct Output {
-        let dataFetched = PassthroughSubject<Void, Never>()
-    }
-    
-    struct NavigationEvent {
-        let backTapped = PassthroughSubject<Void, Never>()
-        let addressSelected = PassthroughSubject<String, Never>()
-    }
-    
-    let output = Output()
-    let navigationEvent = NavigationEvent()
-    
-    let addressService: KakaoAddressServiceProtocol
-    
-    init(addressService: KakaoAddressServiceProtocol) {
-        self.addressService = addressService
-    }
-    
-    func transform(input: AnyPublisher<Input, Never>) -> Output {
-        input
-            .sink { [weak self] event in
-                switch event {
-                case .searchAddress(let query):
-                    self?.handleSearch(query)
-                case .reachedBottom:
-                    self?.handleLoadNextPage()
-                case .addressSelected(let address):
-                    self?.navigationEvent.addressSelected.send(address)
-                case .backTapped:
-                    self?.navigationEvent.backTapped.send()
-                }
-            }
-            .store(in: &cancellables)
-            
-        return output
-    }
-    
-    private func handleSearch(_ query: String) {
-        guard !isFetching else { return }
-        guard !query.isEmpty else {
-            self.searchResults = []
-            self.output.dataFetched.send()
-            return
-        }
-        
-        isFetching = true
-        self.currentQuery = query
-        self.searchCurrentPage = 1
-        
-        addressService.searchAddress(query: query)
-            .sink(receiveCompletion: { [weak self] _ in
-                guard let self else { return }
-                isFetching = false
-            }, receiveValue: { [weak self] response in
-                guard let self else { return }
-                searchResults = response.documents
-                    .filter { $0.address?.region1depthName.contains("서울") ?? false }
-                    .map { $0.addressName }
-                output.dataFetched.send()
-                isFetching = false
-            })
-            .store(in: &cancellables)
-    }
-
-    private func handleLoadNextPage() {
-        self.searchCurrentPage += 1
-    }
-}
 import UIKit
 
 final class AddressSearchViewController: BaseViewController {
